Editing User, making improvements in code

+++ src/main/java/api/mpba/rastvdmy/service/impl/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package api.mpba.rastvdmy.service.impl;
 
-import api.mpba.rastvdmy.config.utils.EncryptionUtil;
-import api.mpba.rastvdmy.dto.request.UserRequest;
+import api.mpba.rastvdmy.dto.request.AdminUpdateUserRequest;
+import api.mpba.rastvdmy.dto.request.UserUpdateRequest;
 import api.mpba.rastvdmy.entity.User;
 import api.mpba.rastvdmy.entity.enums.UserRole;
 import api.mpba.rastvdmy.entity.enums.UserStatus;
@@ -9,8 +9,12 @@ import api.mpba.rastvdmy.exception.ApplicationException;
 import api.mpba.rastvdmy.repository.UserRepository;
 import api.mpba.rastvdmy.service.JwtService;
 import api.mpba.rastvdmy.service.UserService;
-import api.mpba.rastvdmy.service.component.Generator;
+import api.mpba.rastvdmy.service.component.CountryValidator;
+import api.mpba.rastvdmy.service.component.FinancialDataGenerator;
+import api.mpba.rastvdmy.service.component.GenerateAccessToken;
+import api.mpba.rastvdmy.service.component.UserDataValidator;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -22,25 +26,28 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.crypto.SecretKey;
 import java.util.List;
-import java.util.UUID;
 
 //@CacheConfig(cacheNames = {"users"})
 @Service
-public class UserServiceImpl extends Generator implements UserService {
+public class UserServiceImpl extends FinancialDataGenerator implements UserService {
     private final UserRepository userRepository;
-    private final PasswordEncoder encoder;
     private final JwtService jwtService;
+    private final PasswordEncoder passwordEncoder;
+    private final RestTemplate restTemplate;
+    private final GenerateAccessToken generateAccessToken;
 
     @Autowired
     public UserServiceImpl(UserRepository userRepository,
-                           PasswordEncoder encoder,
-                           RestTemplate restTemplate, JwtService jwtService) {
-        super(restTemplate);
+                           JwtService jwtService,
+                           PasswordEncoder passwordEncoder,
+                           RestTemplate restTemplate,
+                           GenerateAccessToken generateAccessToken) {
         this.userRepository = userRepository;
-        this.encoder = encoder;
         this.jwtService = jwtService;
+        this.passwordEncoder = passwordEncoder;
+        this.restTemplate = restTemplate;
+        this.generateAccessToken = generateAccessToken;
     }
 
     public List<User> getUsers() {
@@ -51,8 +58,8 @@ public class UserServiceImpl extends Generator implements UserService {
         return userRepository.findAll(pageable);
     }
 
-    public User getUserById(UUID id) {
-        return userRepository.findById(id.toString())
+    public User getUserByEmail(String email) {
+        return userRepository.findByEmail(email)
                 .orElseThrow(() -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist."));
     }
 
@@ -61,49 +68,74 @@ public class UserServiceImpl extends Generator implements UserService {
                 .orElseThrow(() -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist."));
     }
 
-    private String getUserIdFromToken(HttpServletRequest request) {
+    private String getUserFromToken(HttpServletRequest request) {
         final String token = jwtService.extractToken(request);
-        return jwtService.extractSubject(token); //FIXME: Don't forget that this is users' email!
+        return jwtService.extractSubject(token);
     }
 
-    public User updateUser(HttpServletRequest request, UserRequest userRequest) throws Exception {
-        final String userId = getUserIdFromToken(request);
+    public User updateUser(HttpServletRequest request, UserUpdateRequest userRequest) {
+        final String userEmail = getUserFromToken(request);
 
-        User user = userRepository.findById(userId)
+        User user = userRepository.findByEmail(userEmail)
                 .orElseThrow(() -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist."));
 
         if (user.getStatus() == UserStatus.STATUS_BLOCKED) {
             throw new ApplicationException(HttpStatus.FORBIDDEN, "Operation is forbidden. User is blocked.");
         }
 
-        SecretKey secretKey = EncryptionUtil.getSecretKey();
+        // Validate user data
+        UserDataValidator.isInvalidPassword(userRequest.password());
+        UserDataValidator.isInvalidPhoneNumber(userRequest.phoneNumber());
+        checkIfUserExistsByPhoneNumber(userRequest);
 
-        String encodedEmail = EncryptionUtil.encrypt(userRequest.email(), secretKey, EncryptionUtil.generateIv());
-        String encodedDateOfBirth = EncryptionUtil.encrypt(
-                userRequest.dateOfBirth(), secretKey, EncryptionUtil.generateIv()
-        );
-        String encodedPhoneNumber = EncryptionUtil.encrypt(
-                userRequest.phoneNumber(), secretKey, EncryptionUtil.generateIv()
-        );
+        user.setEmail(userRequest.email());
+        user.setPassword(passwordEncoder.encode(userRequest.password()));
+        user.setPhoneNumber(userRequest.phoneNumber());
 
-        user.setName(userRequest.name());
-        user.setSurname(userRequest.surname());
-        user.setDateOfBirth(encodedDateOfBirth);
-        if (countryExists(userRequest.countryOfOrigin())) {
-            throw new ApplicationException(HttpStatus.BAD_REQUEST, "Country does not exist.");
-        }
-        user.setCountryOrigin(userRequest.countryOfOrigin());
-        user.setEmail(encodedEmail);
-        user.setPassword(userRequest.password());
-        user.encodePassword(encoder);
-        user.setPhoneNumber(encodedPhoneNumber);
         return userRepository.save(user);
     }
 
-    public void uploadUserAvatar(HttpServletRequest request, MultipartFile userAvatar) {
-        final String userId = getUserIdFromToken(request);
+    private void checkIfUserExistsByPhoneNumber(UserUpdateRequest request) {
+        if (userRepository.findByPhoneNumber(request.phoneNumber()).isPresent()) {
+            throw new ApplicationException(
+                    HttpStatus.BAD_REQUEST, "User with phone number " + request.phoneNumber() + " already exists."
+            );
+        }
+    }
 
-        User user = userRepository.findById(userId).orElseThrow(
+    public User updateUserSpecificCredentials(String email, @Valid AdminUpdateUserRequest userRequest) {
+        User user = userRepository.findByEmail(email).orElseThrow(
+                () -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist.")
+        );
+
+        if (user.getStatus() == UserStatus.STATUS_BLOCKED) {
+            throw new ApplicationException(HttpStatus.FORBIDDEN, "Operation is forbidden. User is blocked.");
+        }
+
+        countryValidation(userRequest);
+
+        user.setSurname(userRequest.surname());
+        user.setCountryOrigin(userRequest.countryOfOrigin());
+
+        return userRepository.save(user);
+    }
+
+    public String generateToken(User user) {
+        GenerateAccessToken.TokenDetails generatedToken = generateAccessToken.generate(user);
+        return generatedToken.token();
+    }
+
+    private void countryValidation(AdminUpdateUserRequest request) {
+        CountryValidator countryValidator = new CountryValidator(restTemplate);
+        if (countryValidator.countryExists(request.countryOfOrigin())) {
+            throw new ApplicationException(HttpStatus.BAD_REQUEST, "Country does not exist.");
+        }
+    }
+
+    public void uploadUserAvatar(HttpServletRequest request, MultipartFile userAvatar) {
+        final String userEmail = getUserFromToken(request);
+
+        User user = userRepository.findByEmail(userEmail).orElseThrow(
                 () -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist.")
         );
 
@@ -115,13 +147,14 @@ public class UserServiceImpl extends Generator implements UserService {
             throw new ApplicationException(HttpStatus.BAD_REQUEST, "File must be an image.");
         }
         user.setAvatar(userAvatar.getOriginalFilename());
+
         userRepository.save(user);
     }
 
     public void removeUserAvatar(HttpServletRequest request) {
-        final String userId = getUserIdFromToken(request);
+        final String userEmail = getUserFromToken(request);
 
-        User user = userRepository.findById(userId).orElseThrow(
+        User user = userRepository.findByEmail(userEmail).orElseThrow(
                 () -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist.")
         );
 
@@ -130,16 +163,21 @@ public class UserServiceImpl extends Generator implements UserService {
         }
 
         user.setAvatar("https://i0.wp.com/sbcf.fr/wp-content/uploads/2018/03/sbcf-default-avatar.png?ssl=1");
+
         userRepository.save(user);
     }
 
-    public void updateUserRole(UUID id) {
-        User user = userRepository.findById(id.toString()).orElseThrow(
+    public void updateUserRole(HttpServletRequest request, String email) {
+        User user = userRepository.findByEmail(email).orElseThrow(
                 () -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist.")
         );
 
         if (user.getStatus() == UserStatus.STATUS_BLOCKED) {
             throw new ApplicationException(HttpStatus.FORBIDDEN, "Operation is forbidden. User is blocked.");
+        }
+
+        if (getUserFromToken(request).equals(email)) {
+            throw new ApplicationException(HttpStatus.BAD_REQUEST, "You cannot change your own role.");
         }
 
         switch (user.getRole()) {
@@ -149,10 +187,15 @@ public class UserServiceImpl extends Generator implements UserService {
         userRepository.save(user);
     }
 
-    public void updateUserStatus(UUID id, String status) {
-        User user = userRepository.findById(id.toString()).orElseThrow(
+    public void updateUserStatus(HttpServletRequest request, String email) {
+        User user = userRepository.findByEmail(email).orElseThrow(
                 () -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist.")
         );
+
+        if (getUserFromToken(request).equals(email)) {
+            throw new ApplicationException(HttpStatus.BAD_REQUEST, "You cannot change your own status.");
+        }
+
         switch (user.getStatus()) {
             case STATUS_DEFAULT, STATUS_UNBLOCKED -> {
                 user.setStatus(UserStatus.STATUS_BLOCKED);
@@ -166,8 +209,9 @@ public class UserServiceImpl extends Generator implements UserService {
     }
 
     public void deleteUser(HttpServletRequest request) {
-        final String userId = getUserIdFromToken(request);
-        User user = userRepository.findById(userId).orElseThrow(
+        final String userEmail = getUserFromToken(request);
+
+        User user = userRepository.findByEmail(userEmail).orElseThrow(
                 () -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist.")
         );
 
@@ -180,12 +224,18 @@ public class UserServiceImpl extends Generator implements UserService {
         } else {
             throw new ApplicationException(HttpStatus.BAD_REQUEST, "Make sure to delete all bank accounts first.");
         }
+
     }
 
-    public void deleteUserById(UUID id) {
-        User user = userRepository.findById(id.toString()).orElseThrow(
+    public void deleteUserById(HttpServletRequest request, String email) {
+        User user = userRepository.findByEmail(email).orElseThrow(
                 () -> new ApplicationException(HttpStatus.NOT_FOUND, "User does not exist.")
         );
+
+        if (getUserFromToken(request).equals(email)) {
+            throw new ApplicationException(HttpStatus.BAD_REQUEST, "You cannot delete yourself.");
+        }
+
         if (user.getBankIdentities().isEmpty()) {
             userRepository.delete(user);
         } else {
